refactor: Rename services to UseCase and update package structure for clarity
